pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "specora-core"
version = "0.1.0"
description = "Domain-agnostic Contract-Driven Development engine — builds applications from declarative specifications"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Nicholas Blanchard" }]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "pydantic>=2.0",
    "click>=8.1",
    "rich>=13.0",
    "deepdiff>=7.0",
    "python-dotenv>=1.0",
    "prompt_toolkit>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.4",
]
llm = [
    "openai>=1.0",
    "anthropic>=0.25",
    "httpx>=0.27",
    "pyjwt>=2.8",
]
healer = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "httpx>=0.27",
]
all = [
    "specora-core[dev,llm,healer]",
]

[project.scripts]
specora-core = "forge.cli.main:cli"
spc = "forge.cli.main:cli"
specora-init = "forge.cli.init_project:init_project"

[tool.setuptools.packages.find]
include = ["forge*", "factory*", "engine*", "healer*", "advisor*", "extractor*", "spec*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.meta.yaml", "*.contract.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
